Work with Editing post

+++ Booking.Server/Booking.Domain/Users/User.cs
@@ -11,7 +11,7 @@ public class User : IdentityUser<Guid>
 
 	public string? Avatar { get; set; }
 
-	public int? Rating { get; set; }
+	public float? Rating { get; set; }
 
 	public ICollection<Post>? Posts { get; set; }
 }
